requirements.txt
fastapi
uvicorn
streamlit
pandas
google-genai
requests
pydantic
